+++ oggm/core/flowline.py
@@ -72,13 +72,12 @@ class Flowline(Centerline):
             The water level (to compute volume below sea-level)
         """
 
-        # This is do add flexibility for testing. I have no time for fancier
-        # stuff right now, but I will do this better one day:
+        # This is do add flexibility for testing
         if dx is None:
             dx = 1.
         if line is None:
-            coords = np.arange(0, len(surface_h)-0.5, dx)
-            line = shpg.LineString(np.vstack([coords, coords*0.]).T)
+            coords = np.arange(len(surface_h)) * dx
+            line = shpg.LineString(np.vstack([coords, coords * 0.]).T)
 
         super(Flowline, self).__init__(line, dx, surface_h)
 
@@ -193,7 +192,11 @@ class Flowline(Centerline):
         ds = xr.Dataset()
         ds.coords['x'] = np.arange(nx)
         ds.coords['c'] = [0, 1]
-        ds['linecoords'] = (['x', 'c'], np.asarray(self.line.coords))
+        try:
+            ds['linecoords'] = (['x', 'c'], np.asarray(self.line.coords))
+        except AttributeError:
+            # squeezed lines
+            pass
         ds['surface_h'] = (['x'],  h)
         ds['bed_h'] = (['x'],  self.bed_h)
         ds.attrs['class'] = type(self).__name__